Fixes Heap-buffer-overflow READ in pcpp::HttpRequestFirstLine::parseVersion
Fixes Heap-buffer-overflow READ in pcpp::HttpRequestFirstLine::parseVersion:
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=51006

The `(verPos + 9 - (char*)m_HttpRequest->m_Data)` is `65538` which leads to integer overflow when casted to (uint16_t).
Then `2 > 65535` returns false so it passes the verification.

+++ Packet++/src/HttpLayer.cpp
@@ -368,7 +368,7 @@ void HttpRequestFirstLine::parseVersion()
 	}
 
 	// verify packet doesn't end before the version, meaning still left place for " HTTP/x.y" (9 chars)
-	if ((uint16_t)(verPos + 9 - (char*)m_HttpRequest->m_Data) > m_HttpRequest->getDataLen())
+	if ((verPos + 9 - (char*)m_HttpRequest->m_Data) > m_HttpRequest->getDataLen())
 	{
 		m_Version = HttpVersionUnknown;
 		m_VersionOffset = -1;
